pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "guildbotics"
dynamic = ["version"]
description = "A tool to collaborate with AI agents via a task board"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
  { name = "Ototadana" },
]
keywords = ["agents", "multi-agent", "automation", "workflow"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
  "Intended Audience :: Developers",
  "Environment :: Console",
  "Topic :: Utilities",
]
dependencies = [
  "click>=8.2,<9.0",
  "pydantic>=2.11,<3.0",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "httpx>=0.28,<1.0",
  "cryptography>=45,<47",
  "PyJWT>=2.8",
  "croniter>=3.0",
  "GitPython>=3.1",
  "python-i18n>=0.3",
  "langcodes>=3.3",
  "prompt_toolkit==3.0.51",
  "questionary>=2.1",
  # Engines/brains
  "agno>=1.7,<2.0",
  "requests>=2.32",
  "google-genai>=1.29,<2.0",
  "openai>=1.106,<2.0",
]

[project.urls]
Homepage = "https://github.com/GuildBotics/GuildBotics"
Source = "https://github.com/GuildBotics/GuildBotics"
Issues = "https://github.com/GuildBotics/GuildBotics/issues"

[project.scripts]
guildbotics = "guildbotics.cli:main"

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "types-PyYAML>=6.0",
]
test = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
]
redis = [
  "redis>=5.0",
]


[tool.hatch.build.targets.wheel]
packages = ["guildbotics"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "guildbotics/_version.py"